+++ frontend/PWA_firebase/functions/config/templates/chartofaccounts_template.js
@@ -119,7 +119,7 @@ var coaTemplate = {
        "integers_feild_template_name": {
            "match_mapping_type": "integer",
            "mapping": {
-           "type": "long"
+           "type": "double"
            }
        }
      },
